Ask Optiq: tell a trace without PC samples from a failed read
Every PC-sampling query on a compute schema older than 2.2 is refused as
unsupported, and the controller folded that into a generic error, so
kernel_pc_samples could not tell a trace that cannot hold samples from a
read that went wrong. The three PC-sampling jobs now pass NotSupported
through, the view records it as a layer state of its own, and the
assistant says the trace predates PC sampling. A read that did fail, the
assistant's own or the ISA View's, is reported as a failure rather than as
a capture with no samples, and is retried at most once per call. Only a
read the call actually submitted counts as that retry, and every way a
call ends, clearing the conversation mid-turn included, resets it.

A source file whose lines failed to load was recorded as read, so later
calls skipped it and printed its instructions with no line and no note.
Failures are now kept apart from reads, tried again on the next call, and
named in the answer when a listed instruction sits in that file.

switch_tab re-pinned the turn but let the rest of its batch run with ids
from the trace it had just left. The rest of the batch is now skipped, and
the restart hint goes out whether or not the kind of trace changed.

Kernel min, max, mean and median durations were cut to 32 bits twice, in
the controller's Kernel and again in DataProvider::LoadKernels, so anything
longer than about 4.3 seconds wrapped. Both keep 64 bits now, which also
corrects the compute summary charts.

+++ src/controller/src/compute/rocprofvis_controller_kernel.cpp
@@ -56,25 +56,25 @@ rocprofvis_result_t Kernel::GetUInt64(rocprofvis_property_t property, uint64_t i
             }
             case kRPVControllerKernelDurationMin:
             {
-                *value = (uint64_t)m_duration_min;
+                *value = m_duration_min;
                 result = kRocProfVisResultSuccess;
                 break;
             }
             case kRPVControllerKernelDurationMax:
             {
-                *value = (uint64_t)m_duration_max;
+                *value = m_duration_max;
                 result = kRocProfVisResultSuccess;
                 break;
             }
             case kRPVControllerKernelDurationMedian:
             {
-                *value = (uint64_t)m_duration_median;
+                *value = m_duration_median;
                 result = kRocProfVisResultSuccess;
                 break;
             }
             case kRPVControllerKernelDurationMean:
             {
-                *value = (uint64_t)m_duration_mean;
+                *value = m_duration_mean;
                 result = kRocProfVisResultSuccess;
                 break;
             }
@@ -156,25 +156,25 @@ rocprofvis_result_t Kernel::SetUInt64(rocprofvis_property_t property, uint64_t i
         }
         case kRPVControllerKernelDurationMin:
         {
-            m_duration_min = (uint32_t)value;
+            m_duration_min = value;
             result = kRocProfVisResultSuccess;
             break;
         }
         case kRPVControllerKernelDurationMax:
         {
-            m_duration_max = (uint32_t)value;
+            m_duration_max = value;
             result = kRocProfVisResultSuccess;
             break;
         }
         case kRPVControllerKernelDurationMedian:
         {
-            m_duration_median = (uint32_t)value;
+            m_duration_median = value;
             result = kRocProfVisResultSuccess;
             break;
         }
         case kRPVControllerKernelDurationMean:
         {
-            m_duration_mean = (uint32_t)value;
+            m_duration_mean = value;
             result = kRocProfVisResultSuccess;
             break;
         }
